requirements.txt
PyYAML==5.4.1
opencv-python==4.5.1.48